requirements.txt
python-aliexpress-api
numpy
